+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Encrypt password using bCrypt
 - Validation - add validation
 - Add default role for all new users. 
+- Validation Aspect
+- GlobalException Handler
 
 ## 2026-01-29
 ## Change